HUSH-358 hush-sensor: add support for custom `criSocketPath`
There are cases when snoopy cannot auto-detect CRI socket path.

Add support for specifying a custom path on the Host.

Map the directory of the path. This is done to allow snoopy
to reopen the socket in case it is recreated on the Host.

Moreover, the mounted directory may also contain the containerd socket,
whose path is negotiated over CRI socket. To support this scenario fully
a future commit will add suport for mounting an additional set of
arbitrary paths.

Pass the mounted CRI socket path to snoopy via configuration.