HUSH-361 hush-sensor: change deployment configuration schema in values.yaml
Till now deployment token was specified via root-level `deploymentToken`
string attribute.

Change the schema to be dict-like `deployment.token` in preparation for
making the deployment secret a separate value and adding support for
pre-created secret.

Add "hush-sensor.getDeploymentToken" helper to check that the value was
specified. This value is mandatory. Therefore, fail if it isn't specified.