HUSH-361 hush-sensor: add support for deployment password
Deployment Token was split in the API. The secret part is separate now
and is called Deployment Password.

Add a new input value `deployment.password` to be used in creation of
deployment secret. Note that this value is mandatory now. A future
commit will support providing the password via a pre-created Secret.

Rename the Secret key from `deployment-token` to `deployment-password`
to stress the fact that this is a different value.

Note that Deployment Password isn't base64-encoded (unlike Deployment Token).