HUSH-724 hush-sensor: do not create Namespace in the chart
When a chart creates its Namespace it is impossible to put the Release
object in the same namespace, becauce the namespace doesn't exist yet.

This is a downside because having everything related to some release in
the same namespace makes sense and is probably expected.

Stop creating the Namespace object. Use a pre-created namespace specified
by helm (.Release.Namespace). Allow to override the release namespace by
setting `namespaceOverride` value.

On command line the namespace can be specified as
`helm install -n <namespace-name> ...`.

UI will be updated [1] to use the helm flag accordinglyt.

Note that this change is not backward compatible because it requires
`helm uninstall` first, and then `helm install -n ...`.

This closes HUSH-662 as well.

[1] https://github.com/hushsecurity/mui/pull/25