HUSH-784 hush-sensor: provide snoopy container with SELF_K8S_NAMESPACE env var
Snoopy needs to know the k8s namespace it operates in order not to track
its own processes.

So far this was passed as part of snoopy's config.yaml.

We'd like to keep snoopy config as an interface of items that allows the
admin to configure snoopy's functionalities.
The self_k8s_namespace isn't such item, it is k8s internal information.

Instead, we can pass this as an env var at the container spec using the
downward api.