HUSH-1562 hush-sensor: pass more k8s information into vector env
To be reported in heartbeat events. This will help us better understand
customer deployments and possibly alert on misconfigurations.

The `kube-system` namespace is predefined in k8s clusters, therefore it
is very convenient to use its UID as a stable cluster identifier because:

(a) it is unlikely to be deleted or recreated
(b) Helm provides us the `lookup` function to query Namespace
    attributes, so it can be retrieved easily unlike the real Cluster ID