HUSH-1834 hush-sensor: add known nodes to DaemonSet anti-affinity
This commit adds Azure Virtual Nodes and EKS Fargate to DaemonSet's
anti-affinity rules. It is done in preparation for adding a default
catch-all toleration to our services. Once tolerations allow running on
all nodes we still want to skip some well-known types.

In case of Fargate, DaemonSets are not supported there so avoid
having it stuck in Pending state forever.

In case of Azure Virtual Nodes, we do not want to run there as they are
implemented as containers.

Azure Virtual Node labels are documented at [1]. Specifically, the
nodeSelector for them contains the label 'type: virtual-kubelet'. Add
that label to anti-affinity.

[1] https://github.com/virtual-kubelet/azure-aci#demo