HUSH-2108 hush-sensor: remove usage of `lookup` function
`lookup` is not available in all environments (e.g. `argocd`).

Store the deployment token in k8s Secret next to deployment password
(when the value is given in plaintext). Map it from the Secret
(chart's own or user-created) into all containers.

Stop parsing the token in Helm as we do not always see its value.

Remove helpers that parse the token and all items derived from its
parsed atoms.