HUSH-2108 hush-sensor: allow all pods to query namespaces
Needed to query the UID of `kube-system` which we use as cluster
identifier.

Prep for removing usage of the `lookup` Helm function.

Add a ClusterRole and ServiceAccount to the DaemonSet and Connector
Deployment (they were using the default till now).

Note that only `sentry` has the permission to `list` namespaces
(following minimal required permissions principle).