HUSH-3001 hush-sensor: remove catch-all toleration from Deployments
Leave it on the DaemonSet only.

The reason for this change is that the catch-all toleration allows pods
to run on all nodes including cordoned ones. Running single-instance
Deployments on cordoned nodes is less preferable.

Initially catch-all toleration was added to all deployments to allow
them run on node pools managed/maintained with taints. However, even in
those cases it is assumed that some nodes are untainted. For the
edge case when all nodes are tainted a User has the ability to customize
the tolerations to her needs.