HUSH-3372 makefile: add `--skip-helm-dependencies` argument to `ct`
In preparation for adding `hush-common` library chart which will be a
common dependency for production charts.

The repository will not maintain a `Chart.lock` file because all
library charts will be private.

Without this argument `ct` fails with error [1] because Chart.lock file
doesn't exist.

[1]

Error: the lock file (Chart.lock) is out of sync with the dependencies file (Chart.yaml). Please update the dependencies