HUSH-3973 hush-secretless: add support for ephemeral storage for secretless-manager
There are cases when PV is not defined in a cluster.

Add support for ephemeral storage for secretless-manager so that a PVC
is not required and 'emptyDir' is used instead.

In this mode the manager always boots without having old data on disk.
So, consistency with other instances is dependant on availability of
Hush cloud.