HUSH-4113 hush-secretless: use hush-sensor deployment secret by default
In hush-sensor chart either User provides a pre-populated K8S Secret for
deployment token and password, or they are mandatory parameters and the
chart creates the K8S Secret to use.

In hush-secretless we want to use the same deployment for both
hush-sensor and hush-secretless (for customers that install both).
Moreover, by default we want to assume that hush-sensor is already
installed.

Change the logic in hush-secretless to create K8S Secret only when
token/password are provided explicitly. Otherwise, fallback to
pre-created Secret provided by customer. If both not provided, assume
hush-sensor is installed and use Deployment Secret with the default name
'hush-sensor-deploymentsecret'.